% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sums.R
\name{adds}
\alias{adds}
\title{Adds three numbers}
\usage{
adds(x, y, z)
}
\arguments{
\item{x}{integer}

\item{y}{integer}

\item{z}{integer}
}
\value{
integer
}
\description{
Adds three numbers and returns their sums
}
\examples{
adds(9,7,8)
}
