% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{build_example}
\alias{build_example}
\title{Create example .Rmd file in the current working directory}
\usage{
build_example(lang = c("r", "python"))
}
\description{
Create example .Rmd file in the current working directory
}

