% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hooks.R
\name{go_interactive}
\alias{go_interactive}
\title{Set up your R Markdown document to use DataCamp light}
\usage{
go_interactive(greedy = TRUE)
}
\arguments{
\item{greedy}{whether or not to 'greedily' convert code chunks into DataCamp
Light frames.}
}
\description{
Enable hooks and parsing utilities so that special code chunks where the
\code{ex} and \code{type} options are specified, are converted into
interactive exercises on the resulting HTML page. Use
\code{\link{build_example}} to generate an example document. Simply knitting
this document will generate a document with DataCamp Light powered code
chunks.
}
\details{
With the \code{greedy} argument, you can control which elements of your R
Markdown document are converted into DataCamp Light chunks. By default
\code{greedy} is \code{TRUE}, leading to all R code chunks to be converted to
interactive frames. Only chunks for which you specify the option \code{tut =
FALSE} are not converted. If `greedy` is  \code{FALSE}, only chunks for which
you specify \code{tut = TRUE} are converted.
}

