% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum2.R
\name{sum2}
\alias{sum2}
\title{Function for summing 2 numbers}
\usage{
sum2(num1 = 3, num2 = 4)
}
\arguments{
\item{num1}{First number to sum. Default is 3.}

\item{num2}{Second number to sum. Default is 4.}
}
\value{
Sum of num1 and num2
}
\description{
Function for summing 2 numbers
}
\details{
Function for summing two numbers
}
\examples{
sum2(num1=3, num2=5)
}
