\name{tutorial-package}
\alias{tutorial-package}
\alias{tutorial}
\docType{package}
\title{
\packageTitle{tutorial}
}
\description{
\packageDescription{tutorial}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tutorial}
\packageIndices{tutorial}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{tutorial}

Maintainer: \packageMaintainer{tutorial}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
