% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{build_example}
\alias{build_example}
\title{Create example .Rmd file in the current working directory}
\usage{
build_example()
}
\description{
Create the file \code{example.Rmd} in your current working directory.
This R Markdown files contains example code chunks that can be converted to
an R interactive exercise with the \code{\link{render}} function.
}

