% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderer.R
\name{render}
\alias{render}
\title{Create DataCamp Light readable HTML file}
\usage{
render(input, ...)
}
\arguments{
\item{input}{path to .Rmd file that you want to convert.}

\item{...}{Other arguments that are passed to \code{rmarkdown::render}.}
}
\description{
Currently only supports passing an input file, no other input methods possible.
}
\examples{
\dontrun{
render("my_tutorial.Rmd")
}

}

